refactor: Altera campo options de String para List<String> (JSONB Nativo)
