+++ backend_telemetry/src/main/java/br/com/justina/application/dto/CreateQuestionRequest.java
@@ -1,11 +1,12 @@
 package br.com.justina.application.dto;
 
 import br.com.justina.domain.model.training.QuestionType;
+import java.util.List;
 
 public record CreateQuestionRequest(
         QuestionType type,
         String text,
-        String options, // JSON string
+        List<String> options, // Lista de Strings
         Integer correctIndex,
         String hint,
         String topic,
