+++ backend_telemetry/src/main/java/br/com/justina/application/services/QuestionService.java
@@ -2,7 +2,7 @@ package br.com.justina.application.services;
 
 import br.com.justina.application.dto.CreateQuestionRequest;
 import br.com.justina.domain.model.training.Question;
-import br.com.justina.domain.repository.QuestionRepository;
+import br.com.justina.domain.repository.training.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,11 +16,13 @@ public class QuestionService {
     private final QuestionRepository repository;
 
     public Question create(CreateQuestionRequest request) {
-
         Question question = new Question();
         question.setType(request.type());
         question.setText(request.text());
-        question.setOptions(request.options());
+        
+        // Agora isso funciona, pois request.options() retorna List<String>
+        question.setOptions(request.options()); 
+        
         question.setCorrectIndex(request.correctIndex());
         question.setHint(request.hint());
         question.setTopic(request.topic());
