+++ backend_telemetry/src/main/java/br/com/justina/domain/repository/training/QuestionRepository.java
@@ -1,9 +1,10 @@
-package br.com.justina.domain.repository;
+package br.com.justina.domain.repository.training;
 
 import br.com.justina.domain.model.training.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
-
+import org.springframework.stereotype.Repository;
 import java.util.UUID;
 
+@Repository
 public interface QuestionRepository extends JpaRepository<Question, UUID> {
 }
